Add preload library and initial build system
Create autotools build system.

Add a small LD_PRELOAD library (inspired by path.c from kmod) to wrap the libc
functions that libudev needs for prefixing the paths with $UMOCKDEV_DIR.
Cover sysfs for now (which mostly needs open(), readlink(), and stat()), but
this approach is extensible for intercepting sysctls, device node access, and
similar things in the future as well.

Add a small shell wrapper "umockdev-testbed" which runs the argument under
the appropriate LD_PRELOAD. This is more robust for third party test suites and
also allows us to change the test bed implementation in the future.