Add UMockdevTestbed library
The UMockdevTestbed class is used to build a temporary sysfs file system. You can
add a number of devices including arbitrary sysfs attributes and udev
properties, and then run your software in that test bed that is independent of
the actual hardware it is running on. With this you can simulate particular
hardware in virtual environments up to some degree.