Trace USBFS ioctls
Add interception of open()/ioctl() to the preload library, and log relevant
USBDEVFS ioctls (CONNECTINFO and REAPURB) into a file. Note, most of the
control usbfs ioctls such as CLAIMINTERFACE do not need to be logged as they
do not return device or input dependent results.

Add --ioctl option to umockdump to run a command under ioctl capturing.