Support opening several netlink sockets at the same time
Use a per-socket eventNN unix socket instead of a single one, so that
applications can open multiple netlink sockets and thus udev monitors.

We don't currently have multicast AF_UNIX, and using AF_INET for this would be
harder to implement because we'd need to find a schema for assigning ports, so
just use multiple sockets. We are not going to have a lot of them anyway.