umockdev-run: Propagate signals to child
As we now keep the wrapping umockdev-run process, propagate signals like TERM
or QUIT to the child instead, so that we don't just kill the wrapper.

Also, ensure that we always clean up our testbed even if the child dies with a
signal (raising the signal does not invoke destructors).