Fix detection of mismatching devices when recording into an existing file
Opening the existing file with mode "a" wouldn't allow us to read anything, so
open it with "a+". The read position defaults to the start of the file, so for
ftell() to work we need to seek to the end first. Fix the ignoring of the
first non-comment character and the string comparison.