Fix crash when client sends uevents to closed listeners
We lazily create the unix sockets which mock netlink sockets for
uevents, but so far never cleaned them. It may happen that the client
closes the connection to a previous netlink socket and sends an uevent
after that, so that our server-side socket fails with "Connection
refused". Intercept that and clean up the socket instead of bailing out
with an assertion.

Fixes #53