Avoid buffer overflow in uevent_sender_send()
action, devpath, and other values are unbounded, so ensure that we don't
overflow the property buffer. This does not cross a security boundary
and is not typically used with untrusted data, so not really a
vulnerability. But show a proper error message instead of crashing if it
happens.

Thanks to Simon Schmitt <symenschmitt@web.de> for pointing this out!