Run tests on multiple Ubuntu releases
Ubuntu 14.04 is becoming too old, and misses changes and breakage from
newer system libraries, like issue #69.

Move to building and testing in Launchpad buildd chroots for a given
Ubuntu release and architecture. Add Travis configuration to use this
for the current LTS, the current stable, and the current devel series.