tests/test-umockdev-record.vala: add slow_testbed_factor variable
Some tests expect an operation to happen in less than a specific number
of ms, which is not guaranteed on all testbeds, e.g. slow/overloaded
testbeds. These tests can result in false negative failures.

This adds a multiplier factor to allow slow testbeds to increase the
maximum time that each test expects.