src/umockdev.vala: fix ScriptRunner.op_write
This function reads data into a buffer and compares it with an
expected buffer, with a possible acceptable level of fuzz. However
it does not read it into the temporary buffer correctly nor does it
correctly wait until all data is read into the buffer before comparing
with the expected buffer.