umockdev-run: Don't strip off trailing whitespace from sysfs attributes
libusb 1.0.24 changed the sysfs parser to require attributes be
terminated with a newline [1]. Update umockdev-run to keep trailing
whitespace (in particular, newlines) unmodified, so that the mocked
testbed reproduces them faithfully. Do the minimal changes to the
pre-recorded devices to get along with the new libusb.

Adjust /umockdev-testbed-usb/lsusb accordingly to terminate
busnum/devnum with line feeds.

Fixes #115

[1] https://github.com/libusb/libusb/commit/c3864c822b