tests: Enable NixOS sandboxing
This is more faithful to what happens on real systems, and avoids musl
to leak into the build. This fixes the test-static-code failure, and
reproduces the test failures that happened in the official Nix builds.

Also move from a CLI expression into a standalone file, which makes it
easier to conditionally inject the breakpoint hook.