ioctl-tree: Add function to query size of passed structure
There is no reliable way to know whether the passed argument is a
pointer and what the size of the pointed to structure is. Add a function
to query this information so that external code can make easy use of the
data that is included in the ioctl_tree.c code.