preload: Force a context switch to avoid locking issues
While emulating USB, libusb will busy loop the REAPURB handler (because
poll always returns that the FD is writable). This is not really bad,
but can cause the situation where the looping thread is holding the
ioctl related lock. Other threads making requests which would allow the
replay to continue can end up always waking up with the lock held.

Insert a usleep(0) to force a different thread to be scheduled after
releasing the lock. This is a bit of a workaround, but the incured
overhead is small and it speeds up replay by multiple magnitudes.