ioctl: Load/flush ioctl data implicitly
Rather than requiring explict loads and dirty marking, do this
implicitly. This means that clients only need to resolve pointers
manually and everything else is happening implicitly.

Data is only flushed back when it was modified locally. As such, this
only results in a few more round-trips as e.g. read() data will be
resolved automatically.

If desired, this can be optimized at a later point by improving the
emulation protocol.