meson: Turn "cast-align" back into a warning with werror=false
This is a new option which probably breaks builds a lot on non-x86
architectures. With the new `werror` option we now have a knob to turn
this off, so downgrade to `-W`. With the default `werror=true` it will
still fail to build in CI, as desired; but distros can switch it off.