tests: Work around git misbehaviour
Meson 0.62 turned the "Repository has uncommitted changes" warning into
an error. This happens because `git diff-index HEAD` claims lots of
changes in /source due to the bind mount from the host. Work around that
by adding `--allow-dirty`, which is a good idea anyway for development.

Older versions including 0.61 do not know about `--allow-dirty`, so this
needs to be conditionalized on the version number.

See https://github.com/mesonbuild/meson/issues/10329