umockdev: Use host /sys/fs to allow access to selinux
Newer systemd may warn if it cannot open /sys/fs/selinux/status. As
umockdev is primarily concerned about mocking kernel objects rather than
the content of /sys/fs, an easy solution is to simply not trap /sys/fs
and allow direct access to the file systems mounted there.

This fixes the testsuite on F36 which apparently has a systemd update
that triggers a warning when running udevadm.