ioctl: Make data object usable from python
The bindings cannot deal with a struct object, as it is not a boxed type
and it'll try to ref/unref using g_object_ref. Probably this is
fixable in another way, but this works.

Also, the python bindings cannot modify the byte data directly, so add
two new methods, update and retrieve, to avoid issues.

Please note that this *is* ABI compatible, even if it may not look like
it is. The changes to the header files are restricted to IoctlData, with
the following changes:

* The start of UMockdevIoctlData changes in the following way:
  - GTypeObject parent becomes GObject
  - ref_count is dropped
  - priv pointer is moved to the end
  -> This is OK, as GObject has the exact same memory layout
     (ref_count integer followed by GData* pointer)
  -> The struct size change is OK as the user cannot create it
* The ref/unref functions need to be emulated
* The class changes, but deriving was not useful/possible anyway