HACK: Disable fatal compiler warnings on Gentoo
Between Jan 5 and Jan 8 2024, Gentoo changed something that causes
hundreds of warnings when compiling the vala-generated C code. It's too
hard for me to find out what that was and how to debug it (it's *not* a
new valac version!), so ignore the warnings. It's enough to sort out the
warnings on other OSes, and the unit tests should also ensure that it
actually works.