Skip usbfs_ioctl_pcap on sparc64
Calling Ioctl.USBDEVFS_REAPURB in the test causes a SIGBUS error on
sparc64. It's actually surprising that this test passes on a lot of
architectures -- the pcap file was recorded on x86_64, and there may be
some unaligned access there.

I don't know enough about compilers, alignment, and pcap files to
actually fix this. But the more pressing issue is to fix the package
build on sparc64.

Fixes #218