Formally remove children when the parent device is removed
Although the directory structure for children was being removed when
the parent device was removed, `remove_device()` was not actually being
called on for the children. This meant that (for example) uevents were
not being generated for children when the parent was removed which is
different behavior than on actual hardware.