Handle "connection refused" when sending a uevent
This error was already being handled on the `connect` call right above
these changes, but wasn't being handled on the call to `sendmsg`.
Depending on the timing it was possible for the `connect` call to
succeed while the `sendmsg` call failed with an ECONNREFUSED error
leading to the test crashing from the abort call.